eliminate verify_crl()'s chainOK parameter
Because a malicious CA can create "evil twin" certificates
(certificates that have the same SKI and subject as a valid
certificate), it's possible for a CRL to have a complete certification
path that is invalid even though the CRL is valid (via a path that
isn't yet complete).  Thus, there's no value in distinguishing
complete but invalid paths from incomplete paths.